Replace function checkValidity with reportValidity
While checkValidity and reportValidity's resulting values are alike,
checkValidity will discard the browser's default behavior for invalid
inputs, which displays a message indicating the error to the user.
This would require further customizing error handling on the page.
While this could be useful, researching and training the creation of
custom error behaviors is not something I'm particularly inclined to do
at the moment, with many other priorities at hand.